refactor RentalService
