+++ pom.xml
@@ -133,6 +133,14 @@
             <version>8.0.32</version>
         </dependency>
 
+        <dependency>
+            <groupId>org.projectlombok</groupId>
+            <artifactId>lombok</artifactId>
+            <version>1.18.28</version>
+            <scope>provided</scope>
+        </dependency>
+
+
 
     </dependencies>
