+++ src/main/java/za/ac/cput/controllers/ExceptionHandlerController.java
@@ -1,5 +1,9 @@
 package za.ac.cput.controllers;
-
+/**
+ *
+ * Author: Peter Buckingham (220165289)
+ *
+ */
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
