+++ src/main/java/za/ac/cput/domain/Rental.java
@@ -18,12 +18,17 @@ public class Rental {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private int id;
-    @ManyToOne
+    @ManyToOne //many rentals to one user
     @JoinColumn(name = "user_id")
     private User user;
-    @ManyToOne
+    @ManyToOne //many rentals to one car
     @JoinColumn(name = "car_id")
     private Car car;
+    //driver
+    @ManyToOne //many rentals to one driver
+    @JoinColumn(name = "driver_id")
+    private Driver driver;
+
     private int issuer;
     private int receiver;
     private int fine;
