+++ src/main/java/za/ac/cput/repository/RentalRepository.java
@@ -1,6 +1,7 @@
 package za.ac.cput.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
+import za.ac.cput.domain.Car;
 import za.ac.cput.domain.Rental;
 
 import java.util.List;
@@ -9,15 +10,17 @@ import java.util.Optional;
 /**
  * Peter Buckingham (220169289)
  * Date: March 2023
- * IRentalRepository.java
+ * RentalRepository.java
  */
 
-public interface IRentalRepository extends JpaRepository<Rental, Integer> {
+public interface RentalRepository extends JpaRepository<Rental, Integer> {
 
 
     Optional<Rental> findAllByCarId(int id);
     Optional<Rental> findTopByCarIdOrderByReturnedDateDesc(int id);
 
     List<Rental> findByUserIdAndReturnedDateIsNull(int id);
+
+
 }
