+++ src/main/java/za/ac/cput/repository/UserRepository.java
@@ -1,11 +1,15 @@
 package za.ac.cput.repository;
-
+/**
+ *
+ * Author: Peter Buckingham (220165289)
+ *
+ */
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import za.ac.cput.domain.User;
 
 @Repository
-public interface IUserRepository extends JpaRepository<User, Integer> {
+public interface UserRepository extends JpaRepository<User, Integer> {
     // Add additional methods as we need them
 
 }
