+++ src/main/java/za/ac/cput/service/ICarService.java
@@ -1,5 +1,9 @@
 package za.ac.cput.service;
-
+/**
+ *
+ * Author: Peter Buckingham (220165289)
+ *
+ */
 import za.ac.cput.domain.Car;
 
 import java.util.ArrayList;
@@ -7,11 +11,11 @@ import java.util.ArrayList;
 public interface ICarService extends IService<Car, Integer> {
     Car create(Car car);
 
-    Car read(int id);
+    Car read(Integer id);
 
     Car update(Car car);
 
-    boolean delete(int id);
+    boolean delete(Integer id);
 
     ArrayList<Car> getAll();
 }
