+++ src/main/java/za/ac/cput/service/IRentalService.java
@@ -1,5 +1,9 @@
 package za.ac.cput.service;
-
+/**
+ *
+ * Author: Peter Buckingham (220165289)
+ *
+ */
 import za.ac.cput.domain.Rental;
 
 import java.util.ArrayList;
@@ -7,11 +11,11 @@ import java.util.ArrayList;
 public interface IRentalService extends IService<Rental, Integer> {
     Rental create(Rental rental);
 
-    Rental read(int id);
+    Rental read(Integer id);
 
     Rental update(Rental rental);
 
-    boolean delete(int id);
+    boolean delete(Integer id);
 
     ArrayList<Rental> getAll();
