+++ src/main/java/za/ac/cput/service/IService.java
@@ -1,5 +1,9 @@
 package za.ac.cput.service;
-
+/**
+ *
+ * Author: Peter Buckingham (220165289)
+ *
+ */
 public interface IService<T, ID> {
     T create(T t);
