+++ src/main/java/za/ac/cput/service/IUserService.java
@@ -1,5 +1,9 @@
 package za.ac.cput.service;
-
+/**
+ *
+ * Author: Peter Buckingham (220165289)
+ *
+ */
 import za.ac.cput.domain.User;
 
 import java.util.ArrayList;
@@ -7,11 +11,11 @@ import java.util.ArrayList;
 public interface IUserService extends IService<User, Integer> {
     User create(User user);
 
-    User read(int id);
+    User read(Integer id);
 
     User update(User user);
 
-    boolean delete(int id);
+    boolean delete(Integer id);
 
     ArrayList<User> getAll();
 }
