+++ src/main/java/za/ac/cput/service/impl/UserServiceImpl.java
@@ -1,22 +1,26 @@
 package za.ac.cput.service.impl;
-
+/**
+ *
+ * Author: Peter Buckingham (220165289)
+ *
+ */
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import za.ac.cput.domain.User;
 import za.ac.cput.factory.impl.UserFactory;
-import za.ac.cput.repository.IUserRepository;
+import za.ac.cput.repository.UserRepository;
 import za.ac.cput.service.IUserService;
 
 import java.util.ArrayList;
 import java.util.Optional;
 
 @Service("userServiceImpl")
-public class IUserServiceImpl implements IUserService {
-    private IUserRepository repository = null;
+public class UserServiceImpl implements IUserService {
+    private UserRepository repository = null;
     private UserFactory userFactory;
 
     @Autowired
-    private IUserServiceImpl(IUserRepository repository, UserFactory userFactory) {
+    private UserServiceImpl(UserRepository repository, UserFactory userFactory) {
         this.repository = repository;
         this.userFactory = userFactory;
     }
@@ -28,17 +32,9 @@ public class IUserServiceImpl implements IUserService {
         return repository.save(newUser);
     }
 
-    @Override
-    public User read(Integer integer) {
-        //  Optional<Car> optionalCar = this.repository.findById(id);
-        Optional<User> optionalUser = this.repository.findById(integer);
-        return optionalUser.orElse(null);
-
-
-    }
 
     @Override
-    public User read(int id) {
+    public User read(Integer id) {
         Optional<User> optionalUser = this.repository.findById(id);
         return optionalUser.orElse(null);
     }
@@ -54,14 +50,6 @@ public class IUserServiceImpl implements IUserService {
 
 
     @Override
-    public boolean delete(int id) {
-        if (this.repository.existsById(id)) {
-            this.repository.deleteById(id);
-            return true;
-        }
-        return false;
-    }
-
     public boolean delete(Integer id) {
         if (this.repository.existsById(id)) {
             this.repository.deleteById(id);
@@ -72,7 +60,6 @@ public class IUserServiceImpl implements IUserService {
 
     @Override
     public ArrayList<User> getAll() {
-
         return (ArrayList<User>) this.repository.findAll();
     }
