+++ src/main/resources/application-api.properties
@@ -1,7 +1,7 @@
-spring.security.oauth2.client.registration.google.client-id=371377638852-9r9ubekr3pk082qis0bb93r8b2hafe35.apps.googleusercontent.com
-spring.security.oauth2.client.registration.google.client-secret=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
-spring.security.oauth2.client.registration.google.scope=openid,email,profile
-# This is the main configuration file for the LiberLend system.
+spring.security.oauth2.client.registration.google.client-id=******************************************************
+spring.security.oauth2.client.registration.google.client-secret=**************************************************
+spring.security.oauth2.client.registration.google.scope=***************************************
+
 # It contains the configuration for the database connection
 # Valid values for the database type are:
 #com.mysql.cj.jdbc.Drive
